requirements.txt
asgiref==3.2.10
Django==3.1.2
django-appconf==1.0.4
django-autoslug==1.9.8
django-crispy-forms==1.10.0
django-file-validator==0.0.7
django-imagekit==4.0.2
gunicorn==20.0.4
pilkit==2.0
Pillow==8.1.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
